crypto_currency/requirements.txt
psycopg2-binary
django
schedule
requests
